TINSEL: Discworld 1 PSX music support
Uses FluidSynth v2. `playSEQ()` will dynamically load an in-memory soundfont for each song.

The SoundFont loading is based on the code for the DRAGONS engine, with a fix for the loading the percussion channel instrument.

I've tweaked `playSEQ()` to use some C++11 features, and added a `SmfWriter` struct to make it easier to write SMF commands to the MIDI buffer.

I've changed the base tempo, PSX music doesn't seem to use 109hz timing.

sf2file.cpp: remove faulty logic that limited bank to 127, so we can support percussion channel instruments, which use bank 128.